Initial commit for flattened shapes (no children in shapes)
Box2D supports shapes with multiple children.
Bodies can have multiple fixtures/shapes but each shape could also be composed of smaller parts.
This feature was used only in the implementation of the chain shape.

I believe this 'second layer' in the body hierarchy is unecessarry and imposes a lot of penalties
(memory indirection, more lists, more loops, more complexity) for little gain. The edge shape can
be implemented as a multi-fixture part instead of a multi-shape fixture. With the appropriate helper
functions the burden to the end user should be unnoticeable.

So this commits is the initiation of the effort to remove the concept of multi-part shapes from the library.
This enables further simplifications and performance gains except from those mentioned above.

At the moment the chain shape implementation is not adapted and this will be left for later when the other
parts are stable enough.