Introduce a new interface for broadphase detector
This is the initial commit defining a new (WIP) interface for broadphase detectors.

In the original Box2D broadphase detectors work by using proxies (glue between a fixture and the broadphase)
where the fixture keeps a list of data for all the broadphase detectors it
has been added to. The user has to keep around IDs to access the tree proxies.

In the new model there are no proxies, which in turn means the fixture need not have broadphase-specific code
(plans to generalize the broadphase class later), or keep a list with data for all the broadphases etc.
Also there are no IDs to access the tree nodes.

Instead each broadphase will directly map fixtures to internal nodes (with a hashtable of some sort) and each
fixture will have a (single) precomputed aabb. Overall the new model aims to enable increased efficiency
in the detection (where the bottleneck is) but also will incur some small penalties on structural modification operations.

Note that this commit just contains a new interface with possible future methods to implement, not much of an actual implementation.