Separate the bodies list in static & non-static bodies
The body list in the world class is now split into two logical sections.
The first section has all non-static bodies while the second has only static bodies.
The seperation has no (measurable) performance cost since we just need to store the tail of the doubly-linked list of bodies.
This is useful when we only want to iterate on static/non-static bodies and creates oportunities for optimizations.